refactoring of code
