+++ scripts/01_train_test_split.py
@@ -63,6 +63,13 @@ if __name__ == "__main__":
         default=1,
     )
 
+    parser.add_argument(
+        "--seed",
+        type=int,
+        help="Random seed used to generate the splits. Default 42",
+        default=42
+    )
+
     parser.add_argument("--out_folder", type=str, default=".")
 
     args = parser.parse_args()
@@ -75,7 +82,7 @@ if __name__ == "__main__":
 
     # Splits
     try:
-        cv = StratifiedGroupKFold(n_splits=args.n_splits, shuffle=True, random_state=42)
+        cv = StratifiedGroupKFold(n_splits=args.n_splits, shuffle=True, random_state=args.seed)
 
         # Try to calculate the splits for exception catchment
         next(iter(cv.split(df, df[args.target_col], df[args.group_col])))
