+++ src/taxonomist/user_datasets.py
@@ -1,10 +1,9 @@
-"""
-Defines custom functions for reading dataset data from train-test-splitted csv-files
-"""
 from pathlib import Path
 import pandas as pd
 
-
+"""
+Defines custom functions for reading dataset data from train-test-splitted csv-files
+"""
 def preprocess_dataset(data_folder, dataset_name, csv_path=None, fold=None, label=None):
     data_folder = Path(data_folder)
     fnames = {}
